Broke up lambda code
